Associate images with logged-in user
- ApiUser DTOs now have a list of images that they own
- users can upload any image, but can only view or delete those that
  they own

+++ src/main/java/me/mppombo/synchronyapi/service/ImgurService.java
@@ -1,9 +1,13 @@
 package me.mppombo.synchronyapi.service;
 
-import me.mppombo.synchronyapi.dto.imgur.DeleteOkDto;
+import lombok.RequiredArgsConstructor;
 import me.mppombo.synchronyapi.dto.imgur.PostOkDto;
 import me.mppombo.synchronyapi.dto.imgur.GetOkDto;
+import me.mppombo.synchronyapi.exception.imgur.ImgurDontOwnException;
+import me.mppombo.synchronyapi.exception.imgur.ImgurNotFoundException;
+import me.mppombo.synchronyapi.model.ApiUser;
 import me.mppombo.synchronyapi.model.ImgurImage;
+import me.mppombo.synchronyapi.repository.ApiUserRepository;
 import me.mppombo.synchronyapi.repository.ImgurImageRepository;
 import me.mppombo.synchronyapi.utility.ImgurHelper;
 import org.slf4j.Logger;
@@ -12,28 +16,29 @@ import org.springframework.http.client.MultipartBodyBuilder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.util.List;
 import java.util.Optional;
 
 @Service
+@RequiredArgsConstructor
 public class ImgurService {
     private final Logger logger = LoggerFactory.getLogger(ImgurService.class);
 
     private final ImgurImageRepository imageRepository;
+    private final ApiUserRepository userRepository;
     private final ImgurHelper imgurHelper;
-
-    public ImgurService(ImgurImageRepository imageRepository, ImgurHelper imgurHelper) {
-        this.imageRepository = imageRepository;
-        this.imgurHelper = imgurHelper;
-    }
-
 
     /*
      * Sends a GET request for the Imgur image specified by 'imgHash'.
      * In testing, GET requests to the URI '/images/{hash}' only return either a 200 or 404 response.
      */
-    public ImgurImage getImgurImage(String imgHash) {
+    public ImgurImage getImgurImage(String username, String imgHash) {
 
-        // TODO: ensure requesting user owns image before getting
+        var storedImage = imageRepository
+                .findByImgurId(imgHash)
+                .orElseThrow(ImgurNotFoundException::new);
+        if (!username.equals(storedImage.getOwner().getUsername())) throw new ImgurDontOwnException();
+
         GetOkDto getBody = imgurHelper.getImage(imgHash);
         logger.info("Successful Imgur GET for imgHash='{}'", imgHash);
 
@@ -47,7 +52,7 @@ public class ImgurService {
      * Sends a POST request with a multipart/form-data body which attempts to upload the supplied image to Imgur.
      * Imgur only ever seems to respond with a 200 for uploaded, or a 400 for a malformed request.
      */
-    public ImgurImage uploadImgurImage(MultipartFile image, String title, String description) {
+    public ImgurImage uploadImgurImage(String ownerUsername, MultipartFile image, String title, String description) {
 
         var multipartBodyBuilder = new MultipartBodyBuilder();
         multipartBodyBuilder.part("image", image.getResource());
@@ -58,8 +63,11 @@ public class ImgurService {
         PostOkDto postBody = imgurHelper.postImage(multipartData);
         logger.info("Successful Imgur POST");
 
-        // TODO: associate posted image with logged-in user
+        // Associate this image with the logged-in user. We know they exist, since they couldn't have authenticated
+        // otherwise.
         ImgurImage postedImage = ImgurImage.fromDataDto(postBody.data());
+        ApiUser principalUser = userRepository.findByUsername(ownerUsername).get();
+        postedImage.setOwner(principalUser);
 
         return imageRepository.save(postedImage);
     }
@@ -69,18 +77,20 @@ public class ImgurService {
      * Only ever gets a 200 for successful deletion or a 403 with an "Unauthorized" (even though 403 is technically
      * Forbidden).
      */
-    public boolean deleteImgurImage(String deletehash) {
+    public void deleteImgurImage(String username, String deletehash) {
 
-        // TODO: query logged-in user for deletehash, reject if they don't own it
-        DeleteOkDto deleteBody = imgurHelper.deleteImage(deletehash);
+        var storedImage = imageRepository
+                .findByDeletehash(deletehash)
+                .orElseThrow(ImgurNotFoundException::new);
+        if (!username.equals(storedImage.getOwner().getUsername())) throw new ImgurDontOwnException();
+        imgurHelper.deleteImage(deletehash);
         logger.info("Successful Imgur DELETE for deletehash='{}'", deletehash);
 
-        Optional<ImgurImage> deletedImage = imageRepository.findByDeletehash(deletehash);
-        if (deletedImage.isPresent()) {
-            imageRepository.deleteById(deletedImage.get().getId());
-            logger.info("Deleted {} w/ deletehash='{}' from database", deletedImage, deletehash);
-        }
+        imageRepository.deleteById(storedImage.getId());
+        logger.info("Deleted {} w/ deletehash='{}' from database", storedImage, deletehash);
+    }
 
-        return true;
+    public List<ImgurImage> getAllImagesForUser(ApiUser user) {
+        return imageRepository.findByOwner(user);
     }
 }
